feat(pipelines): Reorder extractor package to separate extractors of content from metadata/file.
